Revoi l'emplacement des fonctions pour Missy
