+++ reverse/client/MissyCore/_model/missyCore.py
@@ -13,11 +13,10 @@ class MissyCore():
         self.listAssignations = []
         self.idAssi = 0
 
-    def initialisation(self, guild: Guild, _kwargs):
-
+    def setup(self, guild: Guild, role: String, channel: String, name: String):
         # set up Target
-        self.channel = guild.get_channel(int(_kwargs["channel"][2:-1]))
-        self.target = Target(1, self.channel, _kwargs["name"], int(_kwargs["role"][3:-1]), guild)
+        self.channel = guild.get_channel(int(channel[2:-1]))
+        self.target = Target(1, self.channel, name, int(role[3:-1]), guild)
         self.listTargets.append(self.target)
 
         # Set up Assignations for Target
@@ -29,37 +28,3 @@ class MissyCore():
         self.listAssignations.append(Assignation(self.idAssi, self.target, "Scribe"))
         self.idAssi = self.idAssi + 1
         self.listAssignations.append(Assignation(self.idAssi, self.target, "Gestionnaire"))
-
-    async def roll(self, date: datetime.date, target: Target):
-        print("--------START------")
-        print(self.listTargets)
-        print("--------------------")
-        print(self.listAssignations)
-        print("---------END--------")
-
-        membersPick = []
-
-        # Pour la target
-        listMembers = target.getAllMembers()
-
-        # Comparer la liste des utilisateurs de ce role sur le serveur 
-        # à la liste des utilisateurs dans la base de données
-        if target.compare(listMembers):
-            for assignation in self.listAssignations:
-                if assignation.target == target:
-
-                    XlastTirage = assignation.getLastXTirages()
-                    users = assignation.clearTirage(listMembers, XlastTirage)
-                    user = assignation.roll(users)
-                    membersPick.append([assignation.name, user])
-
-            print(membersPick)
-            await self.message(membersPick, target, date)
-    
-    async def message(self, membersPick: list, target: Target, date:datetime.date):
-        embed = Embed(title=date, color=0xe80005, timestamp=datetime.datetime.today())
-        
-        for member in membersPick:
-            embed.add_field(name=member[0], value=member[1], inline=False)
-
-        await target.channel.send(content="<@&" + str(target.role.id) + ">",embed=embed)
